factor Shelley client/server from Main
